# Allow reuse of slices exposed through Ready