--- a/raft.go
+++ b/raft.go
@@ -346,7 +346,9 @@ type raft struct {
 	// Messages in this list have the type MsgAppResp, MsgVoteResp, or
 	// MsgPreVoteResp. See the comment in raft.send for details.
 	msgsAfterAppend []pb.Message
-
+	// msgsAfterAppendPool is a pool for msgsAfterAppend to avoid frequently
+	// reallocating the underlying array
+	msgsAfterAppendPool *sync.Pool
 	// the leader id
 	lead uint64
 	// leadTransferee is id of the leader transfer target when its value is not zero.
@@ -424,6 +426,12 @@ func newRaft(c *Config) *raft {
 		preVote:                   c.PreVote,
 		readOnly:                  newReadOnly(c.ReadOnlyOption),
 		disableProposalForwarding: c.DisableProposalForwarding,
+		msgsAfterAppendPool: &sync.Pool{
+			New: func() interface{} {
+				sl := make([]pb.Message, 0, 8)
+				return sl
+			},
+		},
 	}
 
 	cfg, prs, err := confchange.Restore(confchange.Changer{
@@ -545,6 +553,9 @@ func (r *raft) send(m pb.Message) {
 		// because the safety of such behavior has not been formally verified,
 		// we err on the side of safety and omit a `&& !m.Reject` condition
 		// above.
+		if cap(r.msgsAfterAppend) == 0 {
+			r.msgsAfterAppend = r.msgsAfterAppendPool.Get().([]pb.Message)
+		}
 		r.msgsAfterAppend = append(r.msgsAfterAppend, m)
 	} else {
 		if m.To == r.id {