--- a/rawnode.go
+++ b/rawnode.go
@@ -252,7 +252,10 @@ func newStorageAppendMsg(r *raft, rd Ready) pb.Message {
 	// be contained in msgsAfterAppend). This ordering allows the MsgAppResp
 	// handling to use a fast-path in r.raftLog.term() before the newly appended
 	// entries are removed from the unstable log.
-	m.Responses = r.msgsAfterAppend
+
+	//msgsAfterAppend is a pooled slice, so we need to make a deep copy
+	m.Responses = make([]pb.Message, len(r.msgsAfterAppend))
+	copy(m.Responses, r.msgsAfterAppend)
 	if needStorageAppendRespMsg(r, rd) {
 		m.Responses = append(m.Responses, newStorageAppendRespMsg(r, rd))
 	}
@@ -428,6 +431,10 @@ func (rn *RawNode) acceptReady(rd Ready) {
 		}
 	}
 	rn.raft.msgs = nil
+	//release the msgsAfterAppend slice and put it into the msgsAfterAppendPool
+	if cap(rn.raft.msgsAfterAppend) > 0 {
+		rn.raft.msgsAfterAppendPool.Put(rn.raft.msgsAfterAppend[:0])
+	}
 	rn.raft.msgsAfterAppend = nil
 	rn.raft.raftLog.acceptUnstable()
 	if len(rd.CommittedEntries) > 0 {