@CaojiamingAlan 2023-05-12
@pavelkalinnikov Thank you for your comments! I've resolved them.
> we still reallocate and copy the slice.

The trick is at when we put back the slice:
```
rn.raft.msgsAfterAppendPool.Put(rn.raft.msgsAfterAppend[:0])
```
This clears the slice but reserves the underlying array, which we can reuse so that we don't need to allocate a new one each iteration.